Move interrupts and adc out to new modules.
